split taskListUtils into taskListUtils and taskListEntityUtils
